requirements.txt
numpy
RPI.GPIO
scipy
smbus